// src/api/types/rebase_todo_item.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RebaseActionKind } from "./rebase_action_kind";

export type RebaseTodoItem = { action: RebaseActionKind, hash: string, short_hash: string, subject: string, author_name: string, is_published: boolean, reword_message: string | null, };